requirements-dev.txt
pytest>=7.4.0
pytest-cov>=4.1.0
allure-pytest>=2.13.5
pytest-mock>=3.12.0